added c_getslope_ptr test to main.c

+++ xypair.f
@@ -1,4 +1,4 @@
       type, bind(C) :: xypair_t
-            real :: x
-            real :: y
+            real(c_double) :: x
+            real(c_double) :: y
       end type xypair_t
